+++ Sources/KeyboardKit/Actions/KeyboardAction+Actions.swift
@@ -80,6 +80,7 @@ public extension KeyboardAction {
         case .space: { $0?.insertText(.space) }
         case .systemSettings: { $0?.openUrl(.keyboardSettings) }
         case .tab: { $0?.insertText(.tab) }
+        case .text(let text): { $0?.insertText(text) }
         case .url(let url, _): { $0?.openUrl(url) }
         default: nil
         }